+++ src/lib.rs
@@ -58,6 +58,7 @@
 //! └─────────────────────────────────────────────────────────────┘
 //! ```
 
+#![warn(missing_docs)]
 #![deny(
     clippy::unwrap_used,
     clippy::expect_used,